2/Solution-1/requirements.txt
pillow==8.2.0
simplekml